Added Suwa profile and interview reflection
Please see changes, I am unable to allign Suwa's image next to his bio. Ideally I would like it to sit to the right but the container size is too small and there isn't room. I was unable to work out how to change the size. For now it is floating on the left, Nik can you please see how we could format this better?